+++ CommonConcepts/Plugins/Rhetos.Dsl.DefaultConcepts/Computations/FilterInfo.cs
@@ -33,7 +33,6 @@ namespace Rhetos.Dsl.DefaultConcepts
     /// resulting with better performance.
     /// </summary>
     [Export(typeof(IConceptInfo))]
-    [ConceptKeyword("Filter")]
     public class FilterInfo : IConceptInfo
     {
         [ConceptKey]