+++ CommonConcepts/Plugins/Rhetos.Dsl.DefaultConcepts/Computations/FilterPrototypeInfo.cs
@@ -18,20 +18,24 @@
 */
 
 using System.ComponentModel.Composition;
-using System.Linq;
 
 namespace Rhetos.Dsl.DefaultConcepts
 {
     /// <summary>
-    /// A read method that returns a LINQ query for the given parameter value.
-    /// The code snippet should be a lambda expression that returns the query:
-    /// <c>parameter => IQueryable&lt;DataStructureType&gt;</c>.
-    /// The parameter type also represents the filter name.
+    /// Declares an IEnumerable filter on the given data structure.
+    /// The <see cref="FilterInfo.Parameter"/> property is the filter parameter type; it also of represents the filter name.
     /// </summary>
+    /// <remarks>
+    /// This concept generates a partial method 'Filter' on the repository class, in order to specify the method's inputs and outputs.
+    /// This partial method should be implemented by developer on another partial class implementation for the repository class.
+    /// The filter method should return the subset of the given items.
+    /// Note that in many cases it is preferred to use a QueryFilter concept instead;
+    /// it builds a LINQ query and executes the filter directly in database instead of in memory,
+    /// resulting with better performance.
+    /// </remarks>
     [Export(typeof(IConceptInfo))]
-    [ConceptKeyword("Query")]
-    public class QueryExpressionInfo : QueryInfo
+    [ConceptKeyword("Filter")]
+    public class FilterPrototypeInfo : FilterInfo
     {
-        public string QueryImplementation { get; set; }
     }
 }